+++ packages/core/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.31
 
 * Only routes with `on.initial()` or `on.every()` are async. A router with a mix of synchronous and asynchronous routes will match sync routes synchronously.
 * Update to use Hickory's `history.navigate()` for navigation.